-changed light and smoke effect durations back to UINT8s
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@207 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/XML_Explosive.cpp
@@ -187,7 +187,7 @@ explosiveEndElementHandle(void *userData, const char *name)
 		else if(strcmp(name, "ubDuration") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubDuration   = (UINT16) atol(pData->szCharData);
+			pData->curExplosive.ubDuration   = (UINT8) atol(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
